requirements.txt
PyYAML==3.10.0
pika==0.9.13
APScheduler==2.1.1
boto==2.15.0